LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/scale-composition-expected.txt


FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (-0.5) should be [5 1] assert_equals: expected "5 1 " but got "2.5 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0) should be [6 1] assert_equals: expected "6 1 " but got "3 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0.25) should be [6.5 1] assert_equals: expected "6.5 1 " but got "3.25 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0.5) should be [7 1] assert_equals: expected "7 1 " but got "3.5 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (0.75) should be [7.5 1] assert_equals: expected "7.5 1 " but got "3.75 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (1) should be [8 1] assert_equals: expected "8 1 " but got "4 1 "
FAIL Compositing: property <scale> underlying [2 1] from add [3 1] to add [4 1] at (1.5) should be [9 1] assert_equals: expected "9 1 " but got "4.5 1 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (-0.5) should be [2.5 11 22.5] assert_equals: expected "2.5 11 22.5 " but got "2.5 3.5 4.5 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0) should be [4 10 18] assert_equals: expected "4 10 18 " but got "4 5 6 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0.25) should be [4.75 9.5 15.75] assert_equals: expected "4.75 9.5 15.75 " but got "4.75 5.75 6.75 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0.5) should be [5.5 9 13.5] assert_equals: expected "5.5 9 13.5 " but got "5.5 6.5 7.5 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (0.75) should be [6.25 8.5 11.25] assert_equals: expected "6.25 8.5 11.25 " but got "6.25 7.25 8.25 "
PASS Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (1) should be [7 8 9]
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [7 8 9] at (1.5) should be [8.5 7 4.5] assert_equals: expected "8.5 7 4.5 " but got "8.5 9.5 10.5 "
PASS Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (-1) should be [0.5 1]
PASS Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (0) should be [1]
PASS Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (0.25) should be [1.125 1]
PASS Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (0.75) should be [1.375 1]
PASS Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (1) should be [1.5 1]
PASS Compositing: property <scale> underlying [none] from add [none] to replace [1.5 1] at (2) should be [2 1]
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (-1) should be [7 9 11]
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (0) should be [4 5 6]
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (0.25) should be [3.25 4 4.75]
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (0.75) should be [1.75 2 2.25]
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (1) should be [1]
PASS Compositing: property <scale> underlying [none] from add [4 5 6] to replace [none] at (2) should be [-2 -3 -4]
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (-0.5) should be [-2 -1 0] assert_equals: expected "- 2 - 1 0 " but got "- 2 - 2.5 - 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0) should be [1 2 3] assert_equals: expected "1 2 3 " but got "none "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0.25) should be [2.5 3.5 4.5] assert_equals: expected "2.5 3.5 4.5 " but got "2.5 2.75 3 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0.5) should be [4 5 6] assert_equals: expected "4 5 6 " but got "4 4.5 5 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (0.75) should be [5.5 6.5 7.5] assert_equals: expected "5.5 6.5 7.5 " but got "5.5 6.25 7 "
PASS Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (1) should be [7 8 9]
FAIL Compositing: property <scale> underlying [1 2 3] from add [none] to replace [7 8 9] at (1.5) should be [10 11 12] assert_equals: expected "10 11 12 " but got "10 11.5 13 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (-0.5) should be [5.5 14.5 26.5] assert_equals: expected "5.5 14.5 26.5 " but got "5.5 7 8.5 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0) should be [4 10 18] assert_equals: expected "4 10 18 " but got "4 5 6 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0.25) should be [3.25 7.75 13.75] assert_equals: expected "3.25 7.75 13.75 " but got "3.25 4 4.75 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0.5) should be [2.5 5.5 9.5] assert_equals: expected "2.5 5.5 9.5 " but got "2.5 3 3.5 "
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (0.75) should be [1.75 3.25 5.25] assert_equals: expected "1.75 3.25 5.25 " but got "1.75 2 2.25 "
PASS Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (1) should be [1]
FAIL Compositing: property <scale> underlying [1 2 3] from add [4 5 6] to replace [none] at (1.5) should be [-0.5 -3.5 -7.5] assert_equals: expected "- 0.5 - 3.5 - 7.5 " but got "- 0.5 - 1 - 1.5 "
